Tracking: Move plugin export dialog files from workflow to applet/base
